fix signed compare warning

+++ src/impl/compress_zfp.cpp
@@ -246,17 +246,11 @@ private:
       zfp_field_size(field, &sizearray[0]);
       if (zfp_field_type(field) != type ||
           zfp_field_dimensionality(field) != 3 ||
-          sizearray[2] != size[0] || // nz -> slowest varying
-          sizearray[1] != size[1] || // ny
-          sizearray[0] != size[2]) // nx
+          sizearray[2] != static_cast<size_t>(size[0]) || // nz -> slowest varying
+          sizearray[1] != static_cast<size_t>(size[1]) || // ny
+          sizearray[0] != static_cast<size_t>(size[2])) // nx
       {
-        //std::cout << ".type " << (int)zfp_field_type(field)
-        //        << ".dims " << zfp_field_dimensionality(field)
-        //        << ".size " << sizearray[0]
-        //        << ", " << sizearray[1]
-        //        << ", " << sizearray[2]
-        //        << std::endl;
-        throw OpenZGY::Errors::ZgyFormatError("ZFP type or size miamatch");
+          throw OpenZGY::Errors::ZgyFormatError("ZFP type or size miamatch");
       }
 
       const std::int64_t idata_size = size[0]*size[1]*size[2]*sizeof(float);
